fix: fix the display of movies when the user clicks the search button
